test was wrong. corrected it.

+++ float64_float32er_test.go
@@ -10,35 +10,35 @@ func TestFloat64FromFloat32er(t *testing.T) {
 
 	tests := []struct{
 		Value    float32er
-		Expected float64
+		Expected float32
 	}{
 		{
 			Value: testFloat32erNegativeInfinity(),
-			Expected:              math.Inf(-1),
+			Expected:      float32(math.Inf(-1)),
 		},
 		{
 			Value: testFloat32erMin(),
 			Expected:     -math.MaxFloat32,
 		},
 		{
-			Value:     testFloat32erNegativePi(),
-			Expected: float64(float32(-math.Pi)),
+			Value: testFloat32erNegativePi(),
+			Expected:             -math.Pi,
 		},
 		{
-			Value:     testFloat32erNegativeE(),
-			Expected: float64(float32(-math.E)),
+			Value: testFloat32erNegativeE(),
+			Expected:             -math.E,
 		},
 		{
-			Value:     testFloat32erNegativeSqrt2(),
-			Expected: float64(float32(-math.Sqrt2)),
+			Value: testFloat32erNegativeSqrt2(),
+			Expected:             -math.Sqrt2,
 		},
 		{
 			Value: testFloat32erNegativeOne(),
 			Expected:                  -1.0,
 		},
 		{
-			Value:     testFloat32erNegativeLn2(),
-			Expected: float64(float32(-math.Ln2)),
+			Value: testFloat32erNegativeLn2(),
+			Expected:             -math.Ln2,
 		},
 		{
 			Value: testFloat32erNegativeSmallestNonzero(),
@@ -53,39 +53,39 @@ func TestFloat64FromFloat32er(t *testing.T) {
 			Expected:      math.SmallestNonzeroFloat32,
 		},
 		{
-			Value:            testFloat32erLn2(),
-			Expected: float64(float32(math.Ln2)),
+			Value: testFloat32erLn2(),
+			Expected:      math.Ln2,
 		},
 		{
 			Value: testFloat32erOne(),
 			Expected:           1.0,
 		},
 		{
-			Value:            testFloat32erSqrt2(),
-			Expected: float64(float32(math.Sqrt2)),
+			Value: testFloat32erSqrt2(),
+			Expected:      math.Sqrt2,
 		},
 		{
-			Value:            testFloat32erE(),
-			Expected: float64(float32(math.E)),
+			Value: testFloat32erE(),
+			Expected:      math.E,
 		},
 		{
-			Value:            testFloat32erPi(),
-			Expected: float64(float32(math.Pi)),
+			Value: testFloat32erPi(),
+			Expected:      math.Pi,
 		},
 		{
 			Value: testFloat32erMax(),
 			Expected:      math.MaxFloat32,
 		},
 		{
-			Value: testFloat32erInfinity(),
-			Expected:      math.Inf(+1),
+			Value:    testFloat32erInfinity(),
+			Expected: float32(math.Inf(+1)),
 		},
 
 
 
 		{
-			Value: testFloat32erNaN(),
-			Expected:      math.NaN(),
+			Value:    testFloat32erNaN(),
+			Expected: float32(math.NaN()),
 		},
 	}
 
@@ -98,7 +98,7 @@ func TestFloat64FromFloat32er(t *testing.T) {
 			continue
 		}
 
-		y := float64(x)
+		y := float32(x)
 
 		if expected, actual := test.Expected, y; expected != actual {
 			if !(math.IsNaN(float64(expected)) && math.IsNaN(float64(actual))) {
